win32: Fix warnings relating to definition of LONG and use of InterlockedCompareExchange
git-svn-id: svn://rootdirectory.ddns.net/pdclib/branches/shepherd@477 bcf39385-58cc-4174-9fcf-14f50f90dd47

+++ platform/win32/functions/threads/mtx_timedlock.c
@@ -8,7 +8,7 @@ int mtx_timedlock(mtx_t *_PDCLIB_restrict mtx,
 {
     DWORD myId = GetCurrentThreadId();
 
-    if(mtx->_ThreadId == myId) {
+    if(mtx->_ThreadId == (long) myId) {
         mtx->_NestCount++;
         return thrd_success;
     }
